last upgrate, MakeMatrix() works

+++ ConsoleApp1/ConsoleApp1/Point.cs
@@ -8,8 +8,8 @@ namespace ConsoleApp1
 {
     class Point
     {
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float distance { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
+        public double distance { get; set; }
     }
 }
